+++ app/api/v1/bucketlist/__init__.py
@@ -1,0 +1,1 @@
+
