+++ manage.py
@@ -2,19 +2,14 @@ import unittest
 
 from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager, prompt_bool
-
 from app import create_app, db
 from app.api.tests import test_create_user
-from app.api.v1.views import mod as api_views
-from app.site.views import mod as site_views
+
 
 app = create_app('development')
 manager = Manager(app)
 migrate = Migrate(app, db)
 manager.add_command('db', MigrateCommand)
-
-app.register_blueprint(site_views)
-app.register_blueprint(api_views, url_prefix='/api/v1')
 
 
 @manager.command
